Update ilahiHU.txt
History:
This is a very old ilahi from Shaykh Bahauddin Naqshbandi,
founder of the Naqshbandi Tariqa.
This ilahi used to be sung with the zikr Allah in old times.
Nowadays, in the Ansari Qadiri Rifai Tariqa, 
this is one of the main ilahis sung with 
la ila ha illallah.
Refer to Zikr practice section to see how
the ilahi fits very well with this zikr.

Shaykh Abdul Qadir Geylani is the founder
of the Qadiri Tariqa. 
He is also one of the main Pirs of the AQRT Sufi Order.
You can read more about him on the aqrtsufi website.