Update ilahiHU.txt
History:
Seyid Ali Imadaddin Nasimi (Nesimi) 
was a poet in the 14th and 15th century. 
There are many ilahis by him
that are sung in various tekkes
in the world.
You can find some of these ilahis
on this app.

This particular ilahi is very important
in the Ansari Qadiri Rifai Tariqa.
As every ilahi sung in the AQRT zikr, 
it reflects the philosophy of the AQRT. 
This ilahi is usually sung with the zikr
la ila ha illallah.

The first part of the ilahi
speaks about the Nur, Light of Allah
being reflected through the blessed face
of Prophet Muhammad, peace be upon him.

The names of the four main sahaba, companions
of Prophet Muhammad (pbuh)
are mentioned in the following order in this ilahi:
Abu Bakr Siddiq (Abu Bekri - 2nd part, 1st line)
Umar ibn Khattab (Ömer - 2nd part, 2nd line)
Uthman ibn Affan (Osman - 3rd part, 1st line)
Ali ibn Abu Talib (Aliyyel Mürtazadır - 4th part, 2nd line)

By mentioning their names so blissfully in this ilahi, 
Seyyid Nesimi has shown a lot of love and reverence
for all four of them!
By singing this ilahi, especially in our zikr, 
we also show our love, respect and gratefulness
for the four sahaba,
for their contribution to Islam,
for all their sacrifices, dedication and devotion.
Some say, they are like the four pillars of the Ka'aba.
The Prophet, pbuh has mentioned that 
his sahaba are like stars in the sky.
If one follows them, one will be guided
to the right path, to Haqq, Truth.