Update ilahiHU.txt
History:
This ilahi is written by Shaykh Muhyiddin Ansari.
You can read more about him on the aqrtsufi website.
He was a qutb of his time, 
and so was his father, Shaykh Muhammed Ansari.

In this ilahi, we acknowledge 
the founders of the main Tariqas: 
Qadiri, Rifai, Mevlevi, Bektashi, Naqshbandi.
Then the names of the Pirs, founders of the Ansari Qadiri Rifai Tariqa
are saluted:
Shaykh Abdullah Hashimi, 
Shaykh Muhammad Ansari,
Shaykh Muhyiddin Ansari,
Shaykh Taner Ansari,
Shaykha Muzeyyen Ansari.

At the end, the name of the Maydan (main wali) of the area
or main shaykh of the area can be added. e.g. Shaykh Nishaat.
If the name of the Maydan is not known to you, 
you can say Sahibel Maydan (skip the "ya" to fit the tune):
Sahibel Maydan Şeyen lillah illallah Hu