Update poem-1.txt
P: Experience Love
When you adorn yourself with 
patience, kindness and tolerance, 
and then 
understanding and generosity, 
Allah will lower His love 
down on you. 
Then you will experience the love. 
If you are not experiencing this love 
in your life, 
something is wrong 
with your attitude 
and the way 
you are doing things.