Update poem-1.txt
P: Adab
Adab is derived from what pleases Allah 
and what does not please Allah. 
So knowing these things 
and making it a part of you 
and acting with these things 
is the base for adab. 

So adab is actually 
a communication device. 
You are also communicating. 
To whom? 
By interacting with each other, 
you think I am interacting 
with my wife, 
with my husband, 
or friends, this that, 
actually you are interacting 
with the whole universe. 
Everybody is listening to it, 
there is nothing secret, 
whatever you are doing. 
Angels know. 
You remember? 
Allah says 
you have two angels: 
angels write your good deeds 
or 
angels write your bad deeds. 
It means 
they are always with you; 
they know whatever you are doing 
and it is being written.

In Sufism, we try to teach you an adab 
that is pleasing to Allah. 
If you learn and act with it, 
that will get you closer to Allah. 

For example, 
doing your assignment everyday, 
that is an adab. 
Putting Allah in front of anything else 
is a very good adab. 
And treating each other 
for the sake of Allah 
with love 
and compassion 
and forgiveness 
is a very good adab. 
Speaking soft is a good adab. 
Allah likes these things. 

Our Prophet (peace be upon him) 
likes those 
who keep together, 
who do not break hearts 
and help people 
and keep the society positive. 
That is a good adab. 

And these attract blessings of Allah 
through angels and etc. 
Allah creates these angels 
that pray for you until the Judgment Day.