Update poem-1.txt
P: Doing with Understanding and Love
It is important 
to understand 
what you are doing. 
And whatever you are doing 
for Allah, 
do it with an understanding 
and especially with love 
from your heart. 
It has to come from your heart 
whatever you are doing. 
Do everything for Allah's sake 
and then do not regret.