Update poem-1.txt
P: Everything is for love
Now as Sufis, 
we should all focus on Allah 
and do everything for Allah's sake 
and do not expect anything in return 
except we are of course trying to earn Allah's love 
that we want to receive, 
whether Allah gives it or not, 
that is His business 
but we will keep on trying. 
So everything is for love. 
If you take love away 
then there is just business.